+++ graphql-dgs-codegen-core/src/main/kotlin/com/netflix/graphql/dgs/codegen/generators/kotlin/KotlinDataTypeGenerator.kt
@@ -142,52 +142,18 @@ abstract class AbstractKotlinDataTypeGenerator(packageName: String, protected va
         document = document
     )
 
-    object ParserConstants {
-        const val ASSIGNMENT_OPERATOR = " = "
-        const val TYPE = "type"
-        const val NAME = "name"
-        const val REASON = "reason"
-        const val CUSTOM_ANNOTATION = "annotate"
-        const val DEPRECATED = "deprecated"
-        const val INPUTS = "inputs"
-        const val DOT = "."
+    /**
+     * Applies the directives on a field
+     */
+    private fun applyDirectives(directives: List<Directive>, parameterSpec: ParameterSpec.Builder) {
+        parameterSpec.addAnnotations(applyDirectives(directives))
     }
 
     /**
-     * Generates the code block containing the parameters of an annotation in the format key = value
+     * Applies the directives on a graphQL input or type
      */
-    private fun generateCode(value: Value<Value<*>>, prefix: String = "", type: String = ""): CodeBlock =
-        when (value) {
-            is BooleanValue -> CodeBlock.of("$prefix%L", (value as BooleanValue).isValue)
-            is IntValue -> CodeBlock.of("$prefix%L", (value as IntValue).value)
-            is StringValue -> CodeBlock.of("$prefix%S", (value as StringValue).value)
-            is FloatValue -> CodeBlock.of("$prefix%L", (value as FloatValue).value)
-            // If an enum value the prefix/type (key in the parameters map for the enum) is used to get the package name from the config
-            // Limitation: Since it uses the enum key to lookup the package from the configs. 2 enums using different packages cannot have the same keys.
-            is EnumValue -> CodeBlock.of(
-                "$prefix%M",
-                MemberName(
-                    if (prefix.isNotEmpty()) config.includeImports.getOrDefault(prefix.substringBefore(ParserConstants.ASSIGNMENT_OPERATOR), "")
-                    else config.includeImports.getOrDefault(type.substringBefore(ParserConstants.ASSIGNMENT_OPERATOR), ""),
-                    (value as EnumValue).name
-                )
-            )
-            is ArrayValue ->
-                if ((value as ArrayValue).values.isEmpty()) CodeBlock.of("[]")
-                else CodeBlock.of("$prefix[%L]", (value as ArrayValue).values.joinToString { v -> generateCode(value = v, type = if (v is EnumValue) prefix else "").toString() })
-            else -> CodeBlock.of("$prefix%L", value)
-        }
-
-    /**
-     * Parses the inputs argument in the directive to get the input parameters of the annotation
-     */
-    private fun parseInputs(inputs: ObjectValue): List<CodeBlock> {
-        val objectFields: List<ObjectField> = inputs.objectFields
-        return objectFields.fold(mutableListOf(), {
-            codeBlocks, objectField ->
-            codeBlocks.add(generateCode(objectField.value, objectField.name + ParserConstants.ASSIGNMENT_OPERATOR))
-            codeBlocks
-        })
+    private fun applyDirectives(directives: List<Directive>, typeSpec: TypeSpec.Builder) {
+        typeSpec.addAnnotations(applyDirectives(directives))
     }
 
     /**
@@ -209,7 +175,7 @@ abstract class AbstractKotlinDataTypeGenerator(packageName: String, protected va
             annotations, directive ->
             val argumentMap = createArgumentMap(directive)
             if (directive.name == ParserConstants.CUSTOM_ANNOTATION)
-                annotations.add(createCustomAnnotation(argumentMap))
+                annotations.add(customAnnotation(argumentMap, config))
             if (directive.name == ParserConstants.DEPRECATED)
                 if (argumentMap.containsKey(ParserConstants.REASON))
                     annotations.add(deprecatedAnnotation((argumentMap[ParserConstants.REASON] as StringValue).value))
@@ -218,57 +184,6 @@ abstract class AbstractKotlinDataTypeGenerator(packageName: String, protected va
 
             annotations
         })
-    }
-
-    /**
-     * Parses the package value in the directive.
-     * If not present uses the default package in the config for that particular type of annotation.
-     * If neither of them are supplied the package name will be an empty String
-     * Also parses the  simpleName/className from the name argument in the directive
-     */
-    private fun parsePackage(name: String, type: String? = null): Pair<String, String> {
-        var packageName = name.substringBeforeLast(".", "")
-        packageName = if (packageName.isEmpty() && type != null) config.includeImports.getOrDefault(type, "") else packageName
-        return packageName to name.substringAfterLast(".")
-    }
-
-    /**
-     * Creates custom annotation from arguments
-     * name -> Name of the class to be annotated. It will contain className with oor without the package name (Mandatory)
-     * type -> The type of operation intended with this annotation. This value is also used to look up if there is any default packages associated with this annotation in the config
-     * inputs -> These are the input parameters needed for the annotation. If empty no inputs will be present for the annotation
-     */
-    private fun createCustomAnnotation(annotationArgumentMap: MutableMap<String, Value<Value<*>>>): AnnotationSpec {
-        if (annotationArgumentMap.isEmpty() || !annotationArgumentMap.containsKey(ParserConstants.NAME) || annotationArgumentMap[ParserConstants.NAME] is NullValue || (annotationArgumentMap[ParserConstants.NAME] as StringValue).value.isEmpty()) {
-            throw IllegalArgumentException("Invalid annotate directive")
-        }
-        val (packageName, simpleName) = parsePackage(
-            (annotationArgumentMap[ParserConstants.NAME] as StringValue).value,
-            if (annotationArgumentMap.containsKey(ParserConstants.TYPE) && annotationArgumentMap[ParserConstants.TYPE] !is NullValue) (annotationArgumentMap[ParserConstants.TYPE] as StringValue).value else null
-        )
-        val className = ClassName(packageName = packageName, simpleNames = listOf(simpleName))
-        val annotation: AnnotationSpec.Builder = AnnotationSpec.builder(className)
-        if (annotationArgumentMap.containsKey(ParserConstants.INPUTS)) {
-            val codeBlocks: List<CodeBlock> = parseInputs(annotationArgumentMap[ParserConstants.INPUTS] as ObjectValue)
-            codeBlocks.forEach { codeBlock ->
-                annotation.addMember(codeBlock)
-            }
-        }
-        return annotation.build()
-    }
-
-    /**
-     * Applies the directives on a field
-     */
-    private fun applyDirectives(directives: List<Directive>, parameterSpec: ParameterSpec.Builder) {
-        parameterSpec.addAnnotations(applyDirectives(directives))
-    }
-
-    /**
-     * Applies the directives on a graphQL input or type
-     */
-    private fun applyDirectives(directives: List<Directive>, typeSpec: TypeSpec.Builder) {
-        typeSpec.addAnnotations(applyDirectives(directives))
     }
 
     internal fun generate(
